+++ src/_types/checkbox_tree__event_T.d.ts
@@ -1,0 +1,4 @@
+export type checkbox_tree__event_T =
+	CustomEvent<{ currentTarget:EventTarget }>
+	|(Event&{ currentTarget:EventTarget })
+export type checkbox_tree_default_Event_T = checkbox_tree__event_T