+++ src/_types/checkbox_tree_default_Event_T.d.ts
@@ -1,2 +1,0 @@
-export type checkbox_tree_default_Event_T =
-	CustomEvent<{ currentTarget:EventTarget; }>|(Event&{ currentTarget:EventTarget })