+++ src/_types/index.d.ts
@@ -1,1 +1,1 @@
-export * from './checkbox_tree_default_Event_T'
+export * from './checkbox_tree__event_T'