+++ src/all__checkbox__onchange_/index.d.ts
@@ -1,10 +1,16 @@
-import { type WritableAtom_ } from '@ctx-core/nanostores'
-import type { checkbox_tree_default_Event_T } from '../_types'
-export declare function all__checkbox__onchange_<Evt = checkbox_tree_default_Event_T>(
-	store$:WritableAtom_<Record<string, boolean>>,
-	fn:(item_h_selected:Record<string, boolean>, checked:boolean)=>void
-):(event:Evt)=>void;
+import type { WritableAtom_ } from '@ctx-core/nanostores'
+import type { checkbox_tree__event_T } from '../_types'
+export declare function all__checkbox__onchange_<
+	Evt = checkbox_tree__event_T
+>(
+	atom_:WritableAtom_<Record<string, boolean>>,
+	fn:all__checkbox__onchange___fn_T
+):(event:Evt)=>any
 export {
 	all__checkbox__onchange_ as onchange_all_checkbox_,
-	all__checkbox__onchange_ as _onchange_all_checkbox, 
+	all__checkbox__onchange_ as _onchange_all_checkbox,
 }
+export type all__checkbox__onchange___fn_T = (
+	item_R_selected:Record<string, boolean>,
+	checked:boolean
+)=>void