+++ src/checkbox__onchange_/index.d.ts
@@ -1,11 +1,11 @@
-import { type WritableAtom_ } from '@ctx-core/nanostores'
-import { checkbox_tree_default_Event_T } from '../_types'
+import type { WritableAtom_ } from '@ctx-core/nanostores'
+import type { checkbox_tree__event_T } from '../_types'
 export declare function checkbox__onchange_<
-	Evt = checkbox_tree_default_Event_T
+	Evt = checkbox_tree__event_T
 >(
-	store$:WritableAtom_<Record<string, boolean>>
-):(event:Evt, item:string)=>void;
+	atom_:WritableAtom_<Record<string, boolean>>
+):(event:Evt, item:string)=>any
 export {
 	checkbox__onchange_ as onchange_checkbox_,
-	checkbox__onchange_ as _onchange_checkbox, 
+	checkbox__onchange_ as _onchange_checkbox,
 }